Admin gen

+++ database/migrations/2017_01_22_060253_create_plans_table.php
@@ -17,11 +17,11 @@ class CreatePlansTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->text('description')->nullable();
-            $table->double('min_deposit', 15, 8);
-            $table->double('max_deposit', 15, 8);
-            $table->float('sponsor');
-            $table->float('pairing');
-            $table->float('daily');
+            $table->double('min_deposit');
+            $table->double('max_deposit');
+            $table->double('sponsor');
+            $table->double('pairing');
+            $table->double('daily');
             $table->integer('duration');
             $table->string('image')->nullable();
             $table->timestamps();
